fix optimizer problem, we cannot do statistics propagation for those since it might lead to different state types depending on where aggregate is run
