+++ src/include/duckdb/common/types.hpp
@@ -371,12 +371,8 @@ enum class LogicalTypeId : uint8_t {
 
 struct ExtraTypeInfo;
 
-struct LogicalType;
-struct aggregate_state_t {
-	string function_name;
-	vector<LogicalType> bound_argument_types;
-};
 
+struct aggregate_state_t;
 
 struct LogicalType {
 	DUCKDB_API LogicalType();
@@ -612,5 +608,13 @@ bool IsIntegerType() {
 
 bool ApproxEqual(float l, float r);
 bool ApproxEqual(double l, double r);
+
+struct aggregate_state_t {
+	aggregate_state_t(string function_name_p, LogicalType return_type_p, vector<LogicalType> bound_argument_types_p) : function_name(move(function_name_p)), return_type(move(return_type_p)), bound_argument_types(move(bound_argument_types_p)) {}
+
+	string function_name;
+	LogicalType return_type;
+	vector<LogicalType> bound_argument_types;
+};
 
 } // namespace duckdb
